Made gui runnable through maven and jar working with gui
