+++ src/main/java/gui/GuiApplication.java
@@ -1,0 +1,27 @@
+package gui;
+
+import javafx.application.Application;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+import javafx.fxml.FXMLLoader;
+
+import java.io.IOException;
+import java.util.Objects;
+
+public class GuiApplication extends Application {
+    public static void main(String[] args) {
+        launch(args);
+    }
+
+    @Override
+    public void start(Stage stage) throws IOException {
+
+        Parent root = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("/calculator-design.fxml")));
+
+        Scene scene = new Scene(root, 400, 800);
+        stage.setTitle("Calculator App");
+        stage.setScene(scene);
+        stage.show();
+    }
+}
