--- src/main/java/gui/HelloFX.java
@@ -1,49 +1,0 @@
-package gui;
-
-import javafx.application.Application;
-import javafx.event.ActionEvent;
-import javafx.event.EventHandler;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.layout.StackPane;
-import javafx.stage.Stage;
-import javafx.fxml.FXMLLoader;
-
-import java.io.IOException;
-
-public class HelloFX extends Application {
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-    public void start2(Stage primaryStage) {
-        primaryStage.setTitle("Hello World!");
-        Button btn = new Button();
-        btn.setText("Say 'Hello World'");
-        btn.setOnAction(new EventHandler<ActionEvent>() {
-
-            @Override
-            public void handle(ActionEvent event) {
-                System.out.println("Hello World!");
-            }
-        });
-
-        StackPane root = new StackPane();
-        root.getChildren().add(btn);
-        primaryStage.setScene(new Scene(root, 300, 250));
-        primaryStage.show();
-    }
-
-
-    @Override
-    public void start(Stage stage) throws IOException {
-
-        Parent root = FXMLLoader.load(getClass().getResource("/calculator-design.fxml"));
-
-        Scene scene = new Scene(root, 400, 800);
-        stage.setTitle("Calculator App");
-        stage.setScene(scene);
-        stage.show();
-    }
-}
